Removes dependency on kafka server library.
Everything now uses the kafka client library.  This also gets rid of all of the Scala interop which is nice.

A few API changes were necessary to make this work (major version bump) but I believe the end result is a better API.

Also speeds up the tests a little bit by lowering the retry backoff in the default consumer.