Removes spek in favor of JUnit.
Spek is cool, but IntelliJ's integration with it is pretty weak.  The primary problem is that you can't run individual subsets of tests inside IntelliJ which can make debugging failing tests quite painful.

Also, in spek tests the `given` is run once for all `on`s underneath it and the `on` is run once for all `it`s underneath it.  This makes it easy to have tests that influence each other causing unexpected behavior.